Cyans edits
Replacing text from: 'whether you're looking for'
to:  'Mustard is a condiment made from the seeds of a mustard plant.

The whole, ground, cracked, or bruised mustard seeds are mixed with water, vinegar, lemon juice, wine, or other liquids, salt, and often other flavourings and spices, to create a paste or sauce ranging in colour from bright yellow to dark brown. The taste of mustard ranges from sweet to spicy.

Commonly paired with meats and cheeses, mustard is also added to sandwiches, hamburgers, corn dogs, and hot dogs. It is also used as an ingredient in many dressings, glazes, sauces, soups, and marinades. As a cream or as individual seeds, mustard is used as a condiment in the cuisine of India and Bangladesh, the Mediterranean, northern and southeastern Europe, Asia, the Americas, and Africa, making it one of the most popular and widely used spices and condiments in the world.'